% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fits.R
\name{combine_fits}
\alias{combine_fits}
\title{Combine fitted models}
\usage{
combine_fits(models)
}
\arguments{
\item{models}{Returned from \code{\link{fit_models}}.}
}
\value{
list
}
\description{
Combine fitted models
}
