% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fits.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare individual fits to meta analysis fit.}
\usage{
compare_fits(meta_models)
}
\arguments{
\item{meta_models}{Dataframe returned form \code{\link{combine_fits}}.}
}
\value{
Dataframe.
}
\description{
Compare individual fits to meta analysis fit.
}
