% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fits.R
\name{compute_confidence_bands}
\alias{compute_confidence_bands}
\title{Compute confidence bands for the fits}
\usage{
compute_confidence_bands(meta_models)
}
\arguments{
\item{meta_models}{Dataframe.}
}
\value{
Dataframe.
}
\description{
Compute confidence bands for the fits
}
