% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models relating sleep, age and hippocampus volume.}
\usage{
fit_models(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
List of all fitted models.
}
\description{
Fit models relating sleep, age and hippocampus volume.
}
\examples{
#library(metasleep)

}
