% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasleep.R
\docType{package}
\name{metasleep}
\alias{metasleep}
\alias{metasleep-package}
\title{metasleep: A package for additive models relating sleep to hippocampus volume.}
\description{
Intro.
}
\section{Foo functions}{

The foo functions ...
}

